+++ killtracker/management/commands/killtracker_load_eve.py
@@ -2,6 +2,8 @@ import logging
 from django.core.management import call_command
 from django.core.management.base import BaseCommand
 
+from app_utils.logging import LoggerAddTag
+
 from ... import __title__
 from ...constants import (
     EVE_CATEGORY_ID_SHIP,
@@ -10,7 +12,6 @@ from ...constants import (
     EVE_GROUP_MINING_DRONE,
     EVE_GROUP_ORBITAL_INFRASTRUCTURE,
 )
-from ...utils import LoggerAddTag
 
 
 logger = LoggerAddTag(logging.getLogger(__name__), __title__)