+++ killtracker/managers.py
@@ -7,12 +7,13 @@ from django.utils.timezone import now
 
 from allianceauth.services.hooks import get_extension_logger
 
+from app_utils.caching import ObjectCacheMixin
+from app_utils.logging import LoggerAddTag
 from eveuniverse.models import EveEntity
 
 from . import __title__
 from .app_settings import KILLTRACKER_PURGE_KILLMAILS_AFTER_DAYS
 from .core.killmails import Killmail, _KillmailCharacter
-from .utils import LoggerAddTag, ObjectCacheMixin
 
 logger = LoggerAddTag(get_extension_logger(__name__), __title__)